src/requirements.txt
runpod==1.8.1
python-dotenv==1.2.1
openai==2.24.0
orjson==3.10.14
requests==2.32.5